Initial push
Version 1.0.0 has not yet been completed but is mostly done. Numbers can
be read directly from the text box which is useful (Thank God!). The
remainder of the program will be completed most likely tomorrow or the
day after Godwilling with printing answers onto the page through
addition of widgets (or some other means).