Major addition to the README
This explains in detail the incentive and need for this project. Along with what's described it was also a little challenge that I'm enjoying so thank God it's gone so well. I thank Him it's been pretty smooth despite various API fails and failure to understand how to use the API in some cases. But slowly ironing out the flaws and making the code less repetitive and redundant. Currently no speed issues though.