Generic Quotation table completed!
Thank God! The quotation table design is completed and actually generates a word document accordingly. Minor adjustments need to be done in terms of size of columns and general editing. This is the first attempt at this and it seems to operate accurately. Requirements of this are python-docx, a library that helps MS Word manipulation with Python code, and pip, which can be installed by running the get-pip.py file on a Windows/Linux/MacOS computer. All glory to God that this is done!