Aesthetic modification
So far aesthetic modifications have been made to match a full quotation design saving the user lots of time inputting values or even designing the table. The aim is to combine this with a user's Fabric Calculator allowing all data to simply be inputted and in a button press a quotation created. The invoice later on down the line won't be too hard with this already done Godwilling. All glory to God for current progress. And all in one night