Dual document generation
Thank God! So generation of both documents can be correctly done in a single program. Now this needs to be connected with the main Fabric Calculator and fill respective areas with values given. The generated files will ONLY contain the numbers and can be edited and "saved as" accordingly allowing for creation of unique ones each time the program is ran Godwilling